safari-600.7-branch/LayoutTests/sputnik/Conformance/12_Statement/12.7_The_continue_Statement/S12.7_A6-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Cannot use the undeclared label 'LABEL1'.
S12.7_A6

PASS Expected parsing failure

TEST COMPLETE
